requirements.txt
Flask
requests
dotenv
spotipy
google-generativeai
